A new shorter intro
There is a long and proud legacy of drafts that carry introductions concentrating on the
motivations for the last design change rather than the true reasons for
the protocol.  This is my attempt to bring things back on the level.

I didn't stop there though.  I think that we should make this about
HTTP/2 because we don't currently have an extended CONNECT defined for
HTTP/1.1.  This walks back some of the "generic HTTP" language.  More
changes might be needed later in the document for that to be complete,
but I wanted to keep this fix localized.