Improvements to overview
I make a few word-level adjustments, but the goal here was to tweak the
way that this defines key concepts.

I added one concept "WebTransport frames", which isn't necessarily a
global concept, but it might be useful in this document to distinguish
between HTTP/2 frames and the stuff in Section 6.