Fewer headings
The features and protocol details are really different concepts (one
relates to what the protocol does, the other to the implications of
that), but the current structure is poor.  I was taught that you never
leave a section heading empty (S5 currently) and that you never have
just a single subsection (S4.1 and S5.1).  Expanding the scope of the
section is an elegant solution.