Fix the definition of WT_STREAM
The definition was a little loose, so I rewrote it.

We only need to use 0x0a and 0x0b, which is lucky and it lets us avoid
much of the mess that QUIC has around bits in the type.

I've also added definitions for the fields in the frame.